.github/requirements.txt
joblib==1.2.0
numpy==1.23.5
tensorflow_cpu==2.12.0
pandas==1.5.3